+++ CHANGELOG.md
@@ -47,6 +47,10 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## [Unreleased IAVL v1]
 
+## [v0.47.5-v23-osmo-5-iavl-v1](https://github.com/osmosis-labs/cosmos-sdk/releases/tag/v0.47.5-v23-osmo-5-iavl-v1)
+
+* (config) [#572](https://github.com/osmosis-labs/cosmos-sdk/pull/572) Fix module whitelist parsing
+
 ## [v0.47.5-v23-osmo-4-iavl-v1](https://github.com/osmosis-labs/cosmos-sdk/releases/tag/v0.47.5-v23-osmo-4-iavl-v1)
 
 * (config) [#563](https://github.com/osmosis-labs/cosmos-sdk/pull/563) Set lockup as the default whitelisted fast nodes module
@@ -69,6 +73,10 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ## IAVL v23 v0 Releases
 
 ## [Unreleased IAVL v0]
+
+## [v0.47.5-v23-osmo-4](https://github.com/osmosis-labs/cosmos-sdk/releases/tag/v0.47.5-v23-osmo-4)
+
+* (config) [#572](https://github.com/osmosis-labs/cosmos-sdk/pull/572) Fix module whitelist parsing
 
 ## [v0.47.5-v23-osmo-3](https://github.com/osmosis-labs/cosmos-sdk/releases/tag/v0.47.5-v23-osmo-3)